requirements.txt
streamlit
pandas
yfinance
requests
streamlit-autorefresh
python-dateutil